Reorganize Vmmsharp Project Structure
